move metadata model to github folder
