+++ src/github-actions/index.ts
@@ -1,2 +1,1 @@
 export * from "./actions";
-export * from "./metadata-model";
